+++ examples/progress.rs
@@ -54,12 +54,12 @@ pub enum CmdOut {
 }
 
 impl Component for App {
-    type CommandOutput = CmdOut;
     type Init = String;
     type Input = Input;
     type Output = Output;
-    type Root = gtk::Window;
+    type CommandOutput = CmdOut;
     type Widgets = Widgets;
+    type Root = gtk::Window;
 
     fn init_root() -> Self::Root {
         gtk::Window::default()
@@ -100,9 +100,7 @@ impl Component for App {
 
                 append: button = &gtk::Button {
                     set_label: "Compute",
-                    connect_clicked[sender] => move |_| {
-                        sender.input(Input::Compute);
-                    }
+                    connect_clicked => Input::Compute,
                 }
             }
         }