+++ relm4-macros/src/component/mod.rs
@@ -4,11 +4,7 @@ use syn::visit_mut::VisitMut;
 use syn::{parse_quote, Visibility};
 
 use crate::token_streams::{TokenStreams, TraitImplDetails};
-use crate::visitors::{ComponentVisitor, PreAndPostView};
-
-pub(super) mod inject_view_code;
-
-use inject_view_code::inject_view_code;
+use crate::visitors::{ComponentVisitor, PreAndPostView, ViewOutputExpander};
 
 pub(crate) fn generate_tokens(
     vis: &Option<Visibility>,
@@ -17,6 +13,7 @@ pub(crate) fn generate_tokens(
     let mut errors = vec![];
 
     let mut component_visitor = ComponentVisitor::new(&mut errors);
+
     component_visitor.visit_item_impl_mut(&mut component_impl);
 
     let additional_fields = component_visitor.additional_fields.take();
@@ -28,12 +25,20 @@ pub(crate) fn generate_tokens(
 
     let mut struct_fields = None;
 
+    match &component_visitor.view_widgets {
+        None => component_visitor.errors.push(syn::Error::new_spanned(
+            &component_impl,
+            "expected `view!` macro invocation",
+        )),
+        Some(Err(e)) => component_visitor.errors.push(e.clone()),
+        _ => (),
+    }
+
     if let ComponentVisitor {
-        view_widgets: Some(view_widgets),
+        view_widgets: Some(Ok(view_widgets)),
         model_name: Some(model_name),
         root_name: Some(root_name),
         sender_name: Some(sender_name),
-        init: Some(init),
         errors,
         ..
     } = component_visitor
@@ -86,17 +91,14 @@ pub(crate) fn generate_tokens(
             #assign
         };
 
-        let widgets_return_code = quote! {
+        let widgets_return_code = parse_quote! {
             Self::Widgets {
                 #return_fields
                 #additional_fields_return_stream
             }
         };
 
-        let init_injected = match inject_view_code(init, &view_code, &widgets_return_code) {
-            Ok(method) => method,
-            Err(err) => return err.to_compile_error(),
-        };
+        ViewOutputExpander::expand(&mut component_impl, view_code, widgets_return_code, errors);
 
         component_impl.items.push(parse_quote! {
             type Root = #root_widget_type;
@@ -142,10 +144,6 @@ pub(crate) fn generate_tokens(
                 })();
             }
         });
-
-        component_impl
-            .items
-            .push(syn::ImplItem::Method(init_injected));
     }
 
     let outer_attrs = &component_impl.attrs;